Island: Rename cc/resources/auth/auth.py -> cc/resources/auth/authenticate.py
