+++ monkey/monkey_island/cc/resources/auth/authenticate.py
